# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: API Gestionnaire de Bourse ALPE
  version: 1.0.0
  description: |
    API REST pour l'application de gestion des bourses aux vêtements et jouets
    organisées par l'association ALPE Plaisance du Touch.

    ## Authentification
    L'API utilise des tokens JWT (Bearer). Après connexion, incluez le token
    dans l'en-tête `Authorization: Bearer <token>`.

    ## Rôles
    - **deposant** : Gère ses propres listes et articles
    - **benevole** : Scanne et encaisse les ventes
    - **gestionnaire** : Configure les éditions, génère les étiquettes
    - **admin** : Gestion complète (utilisateurs, éditions)

    ## Codes d'erreur
    - 400 : Requête invalide (validation)
    - 401 : Non authentifié
    - 403 : Accès refusé (permissions)
    - 404 : Ressource non trouvée
    - 409 : Conflit (doublon, état invalide)
    - 422 : Entité non traitable (règle métier violée)
    - 429 : Trop de requêtes (rate limiting)
  contact:
    name: ALPE Plaisance du Touch
    email: contact@alpe-plaisance.fr
  license:
    name: Propriétaire
    url: https://alpe-plaisance.fr
servers:
  - url: https://api.bourse-alpe.fr/v1
    description: Production
  - url: https://staging-api.bourse-alpe.fr/v1
    description: Staging
  - url: http://localhost:8000/v1
    description: Développement local
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: Lister les utilisateurs
      description: Réservé aux administrateurs.
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/UserRole'
        - name: search
          in: query
          schema:
            type: string
          description: Recherche par nom ou email
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Liste paginée des utilisateurs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'

    post:
      tags: [Users]
      summary: Créer un utilisateur
      description: Crée un utilisateur (bénévole, gestionnaire). Réservé aux administrateurs.
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Utilisateur créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: Email déjà utilisé
  /users/me:
    get:
      tags: [Users]
      summary: Profil utilisateur courant
      operationId: getCurrentUser
      responses:
        '200':
          description: Profil de l'utilisateur connecté
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    patch:
      tags: [Users]
      summary: Modifier son profil
      operationId: updateCurrentUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSelfUpdate'
      responses:
        '200':
          description: Profil mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{userId}:
    parameters:
      - $ref: '#/components/parameters/userId'
    get:
      tags: [Users]
      summary: Détails d'un utilisateur
      operationId: getUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Détails de l'utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    patch:
      tags: [Users]
      summary: Modifier un utilisateur
      description: Réservé aux administrateurs.
      operationId: updateUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Utilisateur mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    delete:
      tags: [Users]
      summary: Supprimer un utilisateur
      description: Anonymise les données (RGPD). Réservé aux administrateurs.
      operationId: deleteUser
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Utilisateur supprimé/anonymisé

  # ============================================================================
  # INVITATIONS
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    UserRole:
      type: string
      enum: [deposant, benevole, gestionnaire, admin]
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        nom:
          type: string
        prenom:
          type: string
        telephone:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        created_at:
          type: string
          format: date-time
    UserCreate:
      type: object
      required: [email, nom, prenom, role]
      properties:
        email:
          type: string
          format: email
        nom:
          type: string
        prenom:
          type: string
        telephone:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        password:
          type: string
          format: password
    UserUpdate:
      type: object
      properties:
        nom:
          type: string
        prenom:
          type: string
        telephone:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
    UserSelfUpdate:
      type: object
      properties:
        nom:
          type: string
        prenom:
          type: string
        telephone:
          type: string
        password:
          type: string
          format: password
    PaginatedUsers:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer

    # --------------------------------------------------------------------------
    # ÉDITIONS
    # --------------------------------------------------------------------------
...
security:
  - bearerAuth: []